ENH: for training set preparation add option to drop same names without match
For a training set creation, in prepare_name_pairs_pd(), added option to remove all equal names that are
not considered a match. This can happen a lot in actual data, e.g. with franchises that are independent
but do have the same name. So it's a true effect in data, but it screws up our intuitive notion that identical
names should be related. E.g. you may want to set this to true for a model without rank features, which
evaluates string similarity.